# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PhiSphere AI Lab Notebook API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lab-sessions/{id}/upload:
    post:
      operationId: uploadLabSessionFile
      tags: [uploads]
      summary: Upload a CSV or image file to the lab session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /lab-sessions/{id}/load-sample:
    post:
      operationId: loadSampleDataset
      tags: [uploads]
      summary: Load a pre-built sample dataset into a lab session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoadSampleBody"
      responses:
        "200":
          description: Loaded sample result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "404":
          description: Session or dataset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /lab-sessions/{id}/import-openml:
    post:
      operationId: importOpenmlDataset
      tags: [uploads]
      summary: Import a dataset from OpenML into a lab session
      description: Fetches a dataset by ID from the OpenML public repository, converts to CSV if needed, and stores it as session experiment data with attribution.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                datasetId:
                  type: integer
                  description: OpenML dataset ID (e.g. 61 for Iris)
              required:
                - datasetId
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: OpenML API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /openml/search:
    get:
      operationId: searchOpenmlDatasets
      tags: [uploads]
      summary: Search OpenML datasets by name
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search query (dataset name)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Max results (max 25)
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  datasets:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        rows:
                          type: integer
                        features:
                          type: integer
                        format:
                          type: string
                  message:
                    type: string
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CsvColumnStats:
      type: object
      properties:
        min:
          type: number
        max:
          type: number
        mean:
          type: number
      required:
        - min
        - max
        - mean
...
    VisionObject:
      type: object
      properties:
        name:
          type: string
        confidence:
          type: number
      required:
        - name
        - confidence
    VisionAnalysis:
      type: object
      properties:
        available:
          type: boolean
        caption:
          type: string
        captionConfidence:
          type: number
        ocrText:
          type: string
        objects:
          type: array
          items:
            $ref: "#/components/schemas/VisionObject"
        analyzedAt:
          type: string
        error:
          type: string
      required:
        - available
        - analyzedAt
    UploadSummary:
      type: object
      properties:
        columns:
          type: array
          items:
            type: string
        numericColumns:
          type: array
          items:
            type: string
        rowCount:
          type: integer
        stats:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/CsvColumnStats"
        filename:
          type: string
        mimeType:
          type: string
        visionAnalysis:
          $ref: "#/components/schemas/VisionAnalysis"
    UploadResult:
      type: object
      properties:
        uploadType:
          type: string
          enum: [csv, image]
        summary:
          $ref: "#/components/schemas/UploadSummary"
        session:
          $ref: "#/components/schemas/LabSession"
      required:
        - uploadType
        - summary
    LoadSampleBody:
      type: object
      properties:
        datasetId:
          type: string
      required:
        - datasetId
    LabSession:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        domain:
          type: string
        experimentData:
          type: object
          nullable: true
          description: Parsed CSV data or image data attached to the session
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - domain
        - createdAt
        - updatedAt
...